@@ -1,13 +1,13 @@
 import ItemMeta from "./ItemMeta";
 import CommentContainer from "./CommentContainer";
 import React from "react";
-import agent from "../../agent";
 import { connect } from "react-redux";
 import marked from "marked";
 import {
   ITEM_PAGE_LOADED,
   ITEM_PAGE_UNLOADED,
 } from "../../constants/actionTypes";
+import { getItemAndComments } from "./utils/ItemFetcher";
 
 const mapStateToProps = (state) => ({
   ...state.item,
@@ -20,13 +20,11 @@ const mapDispatchToProps = (dispatch) => ({
 });
 
 class Item extends React.Component {
-  componentWillMount() {
-    this.props.onLoad(
-      Promise.all([
-        agent.Items.get(this.props.match.params.id),
-        agent.Comments.forItem(this.props.match.params.id),
-      ])
+  async componentDidMount() {
+    const [item, comments] = await getItemAndComments(
+      this.props.match.params.id
     );
+    this.props.onLoad([item, comments]);
   }
 
   componentWillUnmount() {
@@ -45,7 +43,7 @@ class Item extends React.Component {
       this.props.currentUser &&
       this.props.currentUser.username === this.props.item.seller.username;
     return (
-      <div className="container page">
+      <div className="container page" id="item-container">
         <div className="text-dark">
           <div className="row bg-white p-4">
             <div className="col-6">
@@ -58,7 +56,7 @@ class Item extends React.Component {
             </div>
 
             <div className="col-6">
-              <h1>{this.props.item.title}</h1>
+              <h1 id="card-title">{this.props.item.title}</h1>
               <ItemMeta item={this.props.item} canModify={canModify} />
               <div dangerouslySetInnerHTML={markup}></div>
               {this.props.item.tagList.map((tag) => {
