@@ -3,10 +3,12 @@ import _superagent from "superagent";
 
 const superagent = superagentPromise(_superagent, global.Promise);
 
-const API_ROOT =
+const BACKEND_URL =
   process.env.NODE_ENV !== "production"
-    ? "http://localhost:3000/api"
-    : "https://api.anythink.market/api";
+    ? process.env.REACT_APP_BACKEND_URL
+    : "https://api.anythink.market";
+
+const API_ROOT = `${BACKEND_URL}/api`;
 
 const encode = encodeURIComponent;
 const responseBody = (res) => res.body;
