@@ -2,60 +2,51 @@ import React from "react";
 import { Link } from "react-router-dom";
 import logo from "../imgs/topbar_logo.png";
 
-const LoggedOutView = (props) => {
-  if (!props.currentUser) {
-    return (
-      <ul className="navbar-nav ml-auto">
-        <li className="nav-item">
-          <Link to="/login" className="nav-link">
-            Sign in
-          </Link>
-        </li>
+const LoggedOutView = () => {
+  return (
+    <ul className="navbar-nav ml-auto">
+      <li className="nav-item">
+        <Link to="/login" className="nav-link">
+          Sign in
+        </Link>
+      </li>
 
-        <li className="nav-item">
-          <Link to="/register" className="nav-link">
-            Sign up
-          </Link>
-        </li>
-      </ul>
-    );
-  }
-  return null;
+      <li className="nav-item">
+        <Link to="/register" className="nav-link">
+          Sign up
+        </Link>
+      </li>
+    </ul>
+  );
 };
 
 const LoggedInView = (props) => {
-  if (props.currentUser) {
-    return (
-      <ul className="navbar-nav ml-auto">
-        {props.currentUser.role === "admin" && (
-          <li className="nav-item">
-            <Link to="/editor" className="nav-link">
-              <i className="ion-compose"></i>&nbsp;New Item
-            </Link>
-          </li>
-        )}
+  return (
+    <ul className="navbar-nav ml-auto">
+      <li className="nav-item">
+        <Link to="/editor" className="nav-link">
+          <i className="ion-compose"></i>&nbsp;New Item
+        </Link>
+      </li>
 
-        <li className="nav-item">
-          <Link to="/settings" className="nav-link">
-            <i className="ion-gear-a"></i>&nbsp;Settings
-          </Link>
-        </li>
+      <li className="nav-item">
+        <Link to="/settings" className="nav-link">
+          <i className="ion-gear-a"></i>&nbsp;Settings
+        </Link>
+      </li>
 
-        <li className="nav-item">
-          <Link to={`/@${props.currentUser.username}`} className="nav-link">
-            <img
-              src={props.currentUser.image}
-              className="user-pic pr-1"
-              alt={props.currentUser.username}
-            />
-            {props.currentUser.username}
-          </Link>
-        </li>
-      </ul>
-    );
-  }
-
-  return null;
+      <li className="nav-item">
+        <Link to={`/@${props.currentUser.username}`} className="nav-link">
+          <img
+            src={props.currentUser.image}
+            className="user-pic pr-1"
+            alt={props.currentUser.username}
+          />
+          {props.currentUser.username}
+        </Link>
+      </li>
+    </ul>
+  );
 };
 
 class Header extends React.Component {
@@ -69,9 +60,11 @@ class Header extends React.Component {
           <img alt="logo" src={logo} />
         </Link>
 
-        <LoggedOutView currentUser={this.props.currentUser} />
-
-        <LoggedInView currentUser={this.props.currentUser} />
+        {this.props.currentUser ? (
+          <LoggedInView currentUser={this.props.currentUser} />
+        ) : (
+          <LoggedOutView currentUser={this.props.currentUser} />
+        )}
       </nav>
     );
   }
