@@ -7,7 +7,8 @@ const Banner = () => {
       <div className="container p-4 text-center">
         <img src={logo} alt="banner" />
         <div>
-          <span id="get-part">A place to get</span>
+          <span>A place to </span>
+          <span id="get-part">get</span>
           <span> the cool stuff.</span>
         </div>
       </div>
