@@ -5,9 +5,10 @@ import React from "react";
 const ItemMeta = (props) => {
   const item = props.item;
   return (
-    <div className="d-flex flex-row align-items-center pt-2">
+    <div id="item-meta" className="d-flex flex-row align-items-center pt-2">
       <Link to={`/@${item.seller.username}`}>
         <img
+          id="seller-image"
           src={item.seller.image}
           alt={item.seller.username}
           className="user-pic mr-2"
