@@ -33,6 +33,7 @@ const ItemPreview = (props) => {
     <div
       className="card bg-dark border-light p-3"
       style={{ borderRadius: "20px" }}
+      id={`item_${item.slug}`}
     >
       <img
         alt="item"
@@ -45,7 +46,7 @@ const ItemPreview = (props) => {
           <h3 className="card-title">{item.title}</h3>
           <p className="card-text crop-text-3">{item.description}</p>
         </Link>
-        <div className="d-flex flex-row align-items-center pt-2">
+        <div className="d-flex flex-row align-items-center pt-2 item-footer">
           <Link to={`/@${item.seller.username}`} className="flex-grow-1">
             <img
               src={item.seller.image}
